Improve CC class structure, better base class
