+++ ebcc/cc/base.py
@@ -10,20 +10,27 @@ from ebcc import default_log, init_logging
 from ebcc import numpy as np
 from ebcc import util
 from ebcc.ansatz import Ansatz
+from ebcc.damping import DIIS
+from ebcc.dump import Dump
 from ebcc.logging import ANSI
-from ebcc.precision import types
+from ebcc.precision import cast, types
 
 if TYPE_CHECKING:
-    from logging import Logger
-    from typing import Any, Optional, Union
+    from typing import Any, Callable, Literal, Mapping, Optional, Type, TypeVar, Union
 
-    from pyscf.scf import SCF  # type: ignore
+    from pyscf.scf.hf import SCF  # type: ignore
 
     from ebcc.base import BruecknerEBCC as BaseBrueckner
     from ebcc.base import ERIs as BaseERIs
     from ebcc.base import Fock as BaseFock
+    from ebcc.logging import Logger
     from ebcc.numpy.typing import NDArray  # type: ignore
+    from ebcc.space import Space
     from ebcc.util import Namespace
+
+    ERIsInputType = Union[type[BaseERIs], NDArray[float]]
+    AmplitudeType = TypeVar("AmplitudeType")
+    SpaceType = TypeVar("SpaceType")
 
 
 @dataclass
@@ -72,9 +79,9 @@ class BaseEBCC(ABC):
     Brueckner: type[BaseBrueckner]
 
     # Attributes
-    space: Any
-    amplitudes: Namespace[Any]
-    lambdas: Namespace[Any]
+    space: SpaceType
+    amplitudes: Namespace[AmplitudeType]
+    lambdas: Namespace[AmplitudeType]
 
     def __init__(
         self,
@@ -82,14 +89,14 @@ class BaseEBCC(ABC):
         log: Optional[Logger] = None,
         ansatz: Optional[Union[Ansatz, str]] = "CCSD",
         options: Optional[BaseOptions] = None,
-        space: Optional[Any] = None,
-        omega: Optional[Any] = None,
-        g: Optional[Any] = None,
-        G: Optional[Any] = None,
-        mo_coeff: Optional[Any] = None,
-        mo_occ: Optional[Any] = None,
-        fock: Optional[Any] = None,
-        **kwargs,
+        space: Optional[SpaceType] = None,
+        omega: Optional[NDArray[float]] = None,
+        g: Optional[NDArray[float]] = None,
+        G: Optional[NDArray[float]] = None,
+        mo_coeff: Optional[NDArray[float]] = None,
+        mo_occ: Optional[NDArray[float]] = None,
+        fock: Optional[BaseFock] = None,
+        **kwargs: Any,
     ):
         r"""Initialize the EBCC object.
 
@@ -164,7 +171,7 @@ class BaseEBCC(ABC):
             self.fock = fock
 
         # Attributes:
-        self.e_corr = types[float](0.0)
+        self.e_corr = 0.0
         self.amplitudes = util.Namespace()
         self.converged = False
         self.lambdas = util.Namespace()
@@ -187,22 +194,839 @@ class BaseEBCC(ABC):
         self.log.info(f"{ANSI.B}Space{ANSI.R}: {ANSI.m}{self.space}{ANSI.R}")
         self.log.debug("")
 
-    @abstractmethod
-    @staticmethod
-    def _convert_mf(mf: SCF) -> SCF:
-        """Convert the mean-field object to the appropriate type."""
-        pass
-
-    @abstractmethod
     @property
+    @abstractmethod
     def spin_type(self) -> str:
         """Get a string representation of the spin type."""
         pass
 
-    @abstractmethod
     @property
     def name(self) -> str:
         """Get the name of the method."""
+        return self.spin_type + self.ansatz.name
+
+    def kernel(self, eris: Optional[ERIsInputType] = None) -> float:
+        """Run the coupled cluster calculation.
+
+        Args:
+            eris: Electron repulsion integrals.
+
+        Returns:
+            Correlation energy.
+        """
+        timer = util.Timer()
+
+        # Get the ERIs:
+        eris = self.get_eris(eris)
+
+        # Get the amplitude guesses:
+        amplitudes = self.amplitudes
+        if not amplitudes:
+            amplitudes = self.init_amps(eris=eris)
+
+        # Get the initial energy:
+        e_cc = self.energy(amplitudes=amplitudes, eris=eris)
+
+        self.log.output("Solving for excitation amplitudes.")
+        self.log.debug("")
+        self.log.info(
+            f"{ANSI.B}{'Iter':>4s} {'Energy (corr.)':>16s} {'Energy (tot.)':>18s} "
+            f"{'Δ(Energy)':>13s} {'Δ(Ampl.)':>13s}{ANSI.R}"
+        )
+        self.log.info(f"{0:4d} {e_cc:16.10f} {e_cc + self.mf.e_tot:18.10f}")
+
+        if not self.ansatz.is_one_shot:
+            # Set up DIIS:
+            diis = DIIS()
+            diis.space = self.options.diis_space
+            diis.damping = self.options.damping
+
+            converged = False
+            for niter in range(1, self.options.max_iter + 1):
+                # Update the amplitudes, extrapolate with DIIS and calculate change:
+                amplitudes_prev = amplitudes
+                amplitudes = self.update_amps(amplitudes=amplitudes, eris=eris)
+                vector = self.amplitudes_to_vector(amplitudes)
+                vector = diis.update(vector)
+                amplitudes = self.vector_to_amplitudes(vector)
+                dt = np.linalg.norm(vector - self.amplitudes_to_vector(amplitudes_prev), ord=np.inf)
+
+                # Update the energy and calculate change:
+                e_prev = e_cc
+                e_cc = self.energy(amplitudes=amplitudes, eris=eris)
+                de = abs(e_prev - e_cc)
+
+                # Log the iteration:
+                converged_e = bool(de < self.options.e_tol)
+                converged_t = bool(dt < self.options.t_tol)
+                self.log.info(
+                    f"{niter:4d} {e_cc:16.10f} {e_cc + self.mf.e_tot:18.10f}"
+                    f" {[ANSI.r, ANSI.g][bool(converged_e)]}{de:13.3e}{ANSI.R}"
+                    f" {[ANSI.r, ANSI.g][bool(converged_t)]}{dt:13.3e}{ANSI.R}"
+                )
+
+                # Check for convergence:
+                converged = converged_e and converged_t
+                if converged:
+                    self.log.debug("")
+                    self.log.output(f"{ANSI.g}Converged.{ANSI.R}")
+                    break
+            else:
+                self.log.debug("")
+                self.log.warning(f"{ANSI.r}Failed to converge.{ANSI.R}")
+
+            # Include perturbative correction if required:
+            if self.ansatz.has_perturbative_correction:
+                self.log.debug("")
+                self.log.info("Computing perturbative energy correction.")
+                e_pert = self.energy_perturbative(amplitudes=amplitudes, eris=eris)
+                e_cc += e_pert
+                self.log.info(f"E(pert) = {e_pert:.10f}")
+
+        else:
+            converged = True
+
+        # Update attributes:
+        self.e_corr = e_cc
+        self.amplitudes = amplitudes
+        self.converged = converged
+
+        self.log.debug("")
+        self.log.output(f"E(corr) = {self.e_corr:.10f}")
+        self.log.output(f"E(tot)  = {self.e_tot:.10f}")
+        self.log.debug("")
+        self.log.debug("Time elapsed: %s", timer.format_time(timer()))
+        self.log.debug("")
+
+        return e_cc
+
+    def solve_lambda(self, amplitudes: Optional[Namespace[AmplitudeType]] = None, eris: Optional[ERIsInputType] = None) -> None:
+        """Solve for the lambda amplitudes.
+
+        Args:
+            amplitudes: Cluster amplitudes.
+            eris: Electron repulsion integrals.
+        """
+        timer = util.Timer()
+
+        # Get the ERIs:
+        eris = self.get_eris(eris)
+
+        # Get the amplitudes:
+        amplitudes = self.amplitudes
+        if not amplitudes:
+            amplitudes = self.init_amps(eris=eris)
+
+        # If needed, get the perturbative part of the lambda amplitudes:
+        lambdas_pert = None
+        if self.ansatz.has_perturbative_correction:
+            lambdas_pert = self.update_lams(eris=eris, amplitudes=amplitudes, perturbative=True)
+
+        # Get the initial lambda amplitudes:
+        lambdas = self.lambdas
+        if not lambdas:
+            lambdas = self.init_lams(amplitudes=amplitudes)
+
+        self.log.output("Solving for de-excitation (lambda) amplitudes.")
+        self.log.debug("")
+        self.log.info(f"{ANSI.B}{'Iter':>4s} {'Δ(Ampl.)':>13s}{ANSI.R}")
+
+        # Set up DIIS:
+        diis = DIIS()
+        diis.space = self.options.diis_space
+        diis.damping = self.options.damping
+
+        converged = False
+        for niter in range(1, self.options.max_iter + 1):
+            # Update the lambda amplitudes, extrapolate with DIIS and calculate change:
+            lambdas_prev = lambdas
+            lambdas = self.update_lams(
+                amplitudes=amplitudes,
+                lambdas=lambdas,
+                lambdas_pert=lambdas_pert,
+                eris=eris,
+            )
+            vector = self.lambdas_to_vector(lambdas)
+            vector = diis.update(vector)
+            lambdas = self.vector_to_lambdas(vector)
+            dl = np.linalg.norm(vector - self.lambdas_to_vector(lambdas_prev), ord=np.inf)
+
+            # Log the iteration:
+            converged = bool(dl < self.options.t_tol)
+            self.log.info(f"{niter:4d} {[ANSI.r, ANSI.g][converged]}{dl:13.3e}{ANSI.R}")
+
+            # Check for convergence:
+            if converged:
+                self.log.debug("")
+                self.log.output(f"{ANSI.g}Converged.{ANSI.R}")
+                break
+        else:
+            self.log.debug("")
+            self.log.warning(f"{ANSI.r}Failed to converge.{ANSI.R}")
+
+        self.log.debug("")
+        self.log.debug("Time elapsed: %s", timer.format_time(timer()))
+        self.log.debug("")
+        self.log.debug("")
+
+        # Update attributes:
+        self.lambdas = lambdas
+        self.converged_lambda = converged
+
+    @abstractmethod
+    def ip_eom(self, options: Optional[BaseOptions] = None, **kwargs: Any) -> Any:
+        """Get the IP-EOM object.
+
+        Args:
+            options: Options for the IP-EOM calculation.
+            **kwargs: Additional keyword arguments.
+
+        Returns:
+            IP-EOM object.
+        """
+        pass
+
+    @abstractmethod
+    def ea_eom(self, options: Optional[BaseOptions] = None, **kwargs: Any) -> Any:
+        """Get the EA-EOM object.
+
+        Args:
+            options: Options for the EA-EOM calculation.
+            **kwargs: Additional keyword arguments.
+
+        Returns:
+            EA-EOM object.
+        """
+        pass
+
+    @abstractmethod
+    def ee_eom(self, options: Optional[BaseOptions] = None, **kwargs: Any) -> Any:
+        """Get the EE-EOM object.
+
+        Args:
+            options: Options for the EE-EOM calculation.
+            **kwargs: Additional keyword arguments.
+
+        Returns:
+            EE-EOM object.
+        """
+        pass
+
+    def brueckner(self, *args: Any, **kwargs: Any) -> float:
+        """Run a Brueckner orbital coupled cluster calculation.
+
+        The coupled cluster object will be update in-place.
+
+        Args:
+            *args: Arguments to pass to the Brueckner object.
+            **kwargs: Keyword arguments to pass to the Brueckner object.
+
+        Returns:
+            Correlation energy.
+        """
+        bcc = self.Brueckner(self, *args, **kwargs)
+        return bcc.kernel()
+
+    def write(self, file: str) -> None:
+        """Write the EBCC object to a file.
+
+        Args:
+            file: File to write the object to.
+        """
+        writer = Dump(file)
+        writer.write(self)
+
+    @classmethod
+    def read(cls, file: str, log: Optional[Logger] = None) -> BaseEBCC:
+        """Read the EBCC object from a file.
+
+        Args:
+            file: File to read the object from.
+            log: Logger to use for new object.
+
+        Returns:
+            EBCC object.
+        """
+        reader = Dump(file)
+        return reader.read(cls=cls, log=log)
+
+    @staticmethod
+    @abstractmethod
+    def _convert_mf(mf: SCF) -> SCF:
+        """Convert the mean-field object to the appropriate type."""
+        pass
+
+    def _load_function(
+        self,
+        name: str,
+        eris: Optional[Union[ERIsInputType, Literal[False]]] = False,
+        amplitudes: Optional[Union[Namespace[AmplitudeType], Literal[False]]] = False,
+        lambdas: Optional[Union[Namespace[AmplitudeType], Literal[False]]] = False,
+        **kwargs: Any,
+    ) -> tuple[Callable[..., Any], dict[str, Any]]:
+        """Load a function from the generated code, and return the arguments."""
+        dicts = []
+
+        # Get the ERIs:
+        if not (eris is False):
+            eris = self.get_eris(eris)
+        else:
+            eris = None
+
+        # Get the amplitudes:
+        if not (amplitudes is False):
+            if amplitudes is None:
+                amplitudes = self.amplitudes
+            if amplitudes is None:
+                amplitudes = self.init_amps(eris=eris)
+            dicts.append(dict(amplitudes))
+
+        # Get the lambda amplitudes:
+        if not (lambdas is False):
+            if lambdas is None:
+                lambdas = self.lambdas
+            if lambdas is None:
+                lambdas = self.init_lams(amplitudes=amplitudes if amplitudes else None)
+            dicts.append(dict(lambdas))
+
+        # Get the function:
+        func = getattr(self._eqns, name, None)
+        if func is None:
+            raise util.ModelNotImplemented("%s for %s" % (name, self.name))
+
+        # Get the arguments:
+        if kwargs:
+            dicts.append(kwargs)
+        all_kwargs = self._pack_codegen_kwargs(*dicts, eris=eris)
+
+        return func, all_kwargs
+
+    @abstractmethod
+    def _pack_codegen_kwargs(
+        self, *extra_kwargs: dict[str, Any], eris: Optional[ERIsInputType] = None
+    ) -> dict[str, Any]:
+        """Pack all the keyword arguments for the generated code."""
+        pass
+
+    @abstractmethod
+    def init_amps(self, eris: Optional[ERIsInputType] = None) -> Namespace[AmplitudeType]:
+        """Initialise the cluster amplitudes.
+
+        Args:
+            eris: Electron repulsion integrals.
+
+        Returns:
+            Initial cluster amplitudes.
+        """
+        pass
+
+    @abstractmethod
+    def init_lams(self, amplitude: Optional[Namespace[AmplitudeType]] = None) -> Namespace[AmplitudeType]:
+        """Initialise the cluster lambda amplitudes.
+
+        Args:
+            amplitude: Cluster amplitudes.
+
+        Returns:
+            Initial cluster lambda amplitudes.
+        """
+        pass
+
+    def energy(
+        self,
+        eris: Optional[ERIsInputType] = None,
+        amplitudes: Optional[Namespace[AmplitudeType]] = None,
+    ) -> float:
+        """Calculate the correlation energy.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+
+        Returns:
+            Correlation energy.
+        """
+        func, kwargs = self._load_function(
+            "energy",
+            eris=eris,
+            amplitudes=amplitudes,
+        )
+        return cast(func(**kwargs).real, float)
+
+    def energy_perturbative(
+        self,
+        eris: Optional[ERIsInputType] = None,
+        amplitudes: Optional[Namespace[AmplitudeType]] = None,
+        lambdas: Optional[Namespace[AmplitudeType]] = None,
+    ) -> float:
+        """Calculate the perturbative correction to the correlation energy.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Perturbative energy correction.
+        """
+        func, kwargs = self._load_function(
+            "energy_perturbative",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return cast(func(**kwargs).real, float)
+
+    @abstractmethod
+    def update_amps(
+        self,
+        eris: Optional[ERIsInputType] = None,
+        amplitudes: Optional[Namespace[AmplitudeType]] = None,
+    ) -> Namespace[AmplitudeType]:
+        """Update the cluster amplitudes.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+
+        Returns:
+            Updated cluster amplitudes.
+        """
+        pass
+
+    @abstractmethod
+    def update_lams(
+        self,
+        eris: ERIsInputType = None,
+        amplitude: Optional[Namespace[AmplitudeType]] = None,
+        lambdas: Optional[Namespace[AmplitudeType]] = None,
+        lambdas_pert: Optional[Namespace[AmplitudeType]] = None,
+        perturbative: bool = False,
+    ) -> Namespace[AmplitudeType]:
+        """Update the cluster lambda amplitudes.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitude: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+            lambdas_pert: Perturbative cluster lambda amplitudes.
+            perturbative: Flag to include perturbative correction.
+
+        Returns:
+            Updated cluster lambda amplitudes.
+        """
+        pass
+
+    def make_sing_b_dm(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> Any:
+        r"""Make the single boson density matrix :math:`\langle b \rangle`.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Single boson density matrix.
+        """
+        func, kwargs = self._load_function(
+            "make_sing_b_dm",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    def make_rdm1_b(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None, unshifted: bool = True, hermitise: bool = True) -> Any:
+        r"""Make the one-particle boson reduced density matrix :math:`\langle b^+ c \rangle`.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+            unshifted: If `self.options.shift` is `True`, return the unshifted density matrix. Has
+                no effect if `self.options.shift` is `False`.
+            hermitise: Hermitise the density matrix.
+
+        Returns:
+            One-particle boson reduced density matrix.
+        """
+        func, kwargs = self._load_function(
+            "make_rdm1_b",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        dm = func(**kwargs)
+
+        if hermitise:
+            dm = 0.5 * (dm + dm.T)
+
+        if unshifted and self.options.shift:
+            dm_cre, dm_ann = self.make_sing_b_dm()
+            xi = self.xi
+            dm[np.diag_indices_from(dm)] -= xi * (dm_cre + dm_ann) - xi**2
+
+        return dm
+
+    @abstractmethod
+    def make_rdm1_f(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None, hermitise: bool = True) -> Any:
+        r"""Make the one-particle fermionic reduced density matrix :math:`\langle i^+ j \rangle`.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+            hermitise: Hermitise the density matrix.
+
+        Returns:
+            One-particle fermion reduced density matrix.
+        """
+        pass
+
+    @abstractmethod
+    def make_rdm2_f(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None, hermitise: bool = True) -> Any:
+        r"""Make the two-particle fermionic reduced density matrix :math:`\langle i^+j^+lk \rangle`.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+            hermitise: Hermitise the density matrix.
+
+        Returns:
+            Two-particle fermion reduced density matrix.
+        """
+        pass
+
+    @abstractmethod
+    def make_eb_coup_rdm(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None, unshifted: bool = True, hermitise: bool = True) -> Any:
+        r"""Make the electron-boson coupling reduced density matrix :math:`\langle b^+ c^+ c b \rangle`.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+            unshifted: If `self.options.shift` is `True`, return the unshifted density matrix. Has
+                no effect if `self.options.shift` is `False`.
+            hermitise: Hermitise the density matrix.
+
+        Returns:
+            Electron-boson coupling reduced density matrix.
+        """
+        pass
+
+    def hbar_matvec_ip(self, r1: AmplitudeType, r2: AmplitudeType, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType]:
+        """Compute the product between a state vector and the IP-EOM Hamiltonian.
+
+        Args:
+            r1: State vector (single excitations).
+            r2: State vector (double excitations).
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+
+        Returns:
+            Products between the state vectors and the IP-EOM Hamiltonian for the singles and
+            doubles.
+        """
+        func, kwargs = self._load_function(
+            "hbar_matvec_ip",
+            eris=eris,
+            amplitudes=amplitudes,
+            r1=r1,
+            r2=r2,
+        )
+        return func(**kwargs)
+
+    def hbar_matvec_ea(self, r1: AmplitudeType, r2: AmplitudeType, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, AmplitudeType]:
+        """Compute the product between a state vector and the EA-EOM Hamiltonian.
+
+        Args:
+            r1: State vector (single excitations).
+            r2: State vector (double excitations).
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+
+        Returns:
+            Products between the state vectors and the EA-EOM Hamiltonian for the singles and
+            doubles.
+        """
+        func, kwargs = self._load_function(
+            "hbar_matvec_ea",
+            eris=eris,
+            amplitudes=amplitudes,
+            r1=r1,
+            r2=r2,
+        )
+        return func(**kwargs)
+
+    def hbar_matvec_ee(self, r1: AmplitudeType, r2: AmplitudeType, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, AmplitudeType]:
+        """Compute the product between a state vector and the EE-EOM Hamiltonian.
+
+        Args:
+            r1: State vector (single excitations).
+            r2: State vector (double excitations).
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+
+        Returns:
+            Products between the state vectors and the EE-EOM Hamiltonian for the singles and
+            doubles.
+        """
+        func, kwargs = self._load_function(
+            "hbar_matvec_ee",
+            eris=eris,
+            amplitudes=amplitudes,
+            r1=r1,
+            r2=r2,
+        )
+        return func(**kwargs)
+
+    def make_ip_mom_bras(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, ...]:
+        """Get the bra vectors to construct IP-EOM moments.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Bra vectors for IP-EOM moments.
+        """
+        func, kwargs = self._load_function(
+            "make_ip_mom_bras",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    def make_ea_mom_bras(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, ...]:
+        """Get the bra vectors to construct EA-EOM moments.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Bra vectors for EA-EOM moments.
+        """
+        func, kwargs = self._load_function(
+            "make_ea_mom_bras",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    def make_ee_mom_bras(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, ...]:
+        """Get the bra vectors to construct EE-EOM moments.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Bra vectors for EE-EOM moments.
+        """
+        func, kwargs = self._load_function(
+            "make_ee_mom_bras",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    def make_ip_mom_kets(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, ...]:
+        """Get the ket vectors to construct IP-EOM moments.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Ket vectors for IP-EOM moments.
+        """
+        func, kwargs = self._load_function(
+            "make_ip_mom_kets",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    def make_ea_mom_kets(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, ...]:
+        """Get the ket vectors to construct EA-EOM moments.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Ket vectors for EA-EOM moments.
+        """
+        func, kwargs = self._load_function(
+            "make_ea_mom_kets",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    def make_ee_mom_kets(self, eris: Optional[ERIsInputType] = None, amplitudes: Optional[Namespace[AmplitudeType]] = None, lambdas: Optional[Namespace[AmplitudeType]] = None) -> tuple[AmplitudeType, ...]:
+        """Get the ket vectors to construct EE-EOM moments.
+
+        Args:
+            eris: Electron repulsion integrals.
+            amplitudes: Cluster amplitudes.
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Ket vectors for EE-EOM moments.
+        """
+        func, kwargs = self._load_function(
+            "make_ee_mom_kets",
+            eris=eris,
+            amplitudes=amplitudes,
+            lambdas=lambdas,
+        )
+        return func(**kwargs)
+
+    @abstractmethod
+    def energy_sum(self, *args: str, signs_dict: Optional[dict[str, int]] = None) -> NDArray[float]:
+        """Get a direct sum of energies.
+
+        Args:
+            *args: Energies to sum. Subclass should specify a subscript, and optionally spins.
+            signs_dict: Signs of the energies in the sum. Default sets `("o", "O", "i")` to be
+                positive, and `("v", "V", "a", "b")` to be negative.
+
+        Returns:
+            Sum of energies.
+        """
+        pass
+
+    @abstractmethod
+    def amplitudes_to_vector(self, amplitudes: Namespace[AmplitudeType]) -> NDArray[float]:
+        """Construct a vector containing all of the amplitudes used in the given ansatz.
+
+        Args:
+            amplitudes: Cluster amplitudes.
+
+        Returns:
+            Cluster amplitudes as a vector.
+        """
+        pass
+
+    @abstractmethod
+    def vector_to_amplitudes(self, vector: NDArray[float]) -> Namespace[AmplitudeType]:
+        """Construct a namespace of amplitudes from a vector.
+
+        Args:
+            vector: Cluster amplitudes as a vector.
+
+        Returns:
+            Cluster amplitudes.
+        """
+        pass
+
+    @abstractmethod
+    def lambdas_to_vector(self, lambdas: Namespace[AmplitudeType]) -> NDArray[float]:
+        """Construct a vector containing all of the lambda amplitudes used in the given ansatz.
+
+        Args:
+            lambdas: Cluster lambda amplitudes.
+
+        Returns:
+            Cluster lambda amplitudes as a vector.
+        """
+        pass
+
+    @abstractmethod
+    def vector_to_lambdas(self, vector: NDArray[float]) -> Namespace[AmplitudeType]:
+        """Construct a namespace of lambda amplitudes from a vector.
+
+        Args:
+            vector: Cluster lambda amplitudes as a vector.
+
+        Returns:
+            Cluster lambda amplitudes.
+        """
+        pass
+
+    @abstractmethod
+    def excitations_to_vector_ip(self, *excitations: Namespace[AmplitudeType]) -> NDArray[float]:
+        """Construct a vector containing all of the IP-EOM excitations.
+
+        Args:
+            excitations: IP-EOM excitations.
+
+        Returns:
+            IP-EOM excitations as a vector.
+        """
+        pass
+
+    @abstractmethod
+    def excitations_to_vector_ea(self, *excitations: Namespace[AmplitudeType]) -> NDArray[float]:
+        """Construct a vector containing all of the EA-EOM excitations.
+
+        Args:
+            excitations: EA-EOM excitations.
+
+        Returns:
+            EA-EOM excitations as a vector.
+        """
+        pass
+
+    @abstractmethod
+    def excitations_to_vector_ee(self, *excitations: Namespace[AmplitudeType]) -> NDArray[float]:
+        """Construct a vector containing all of the EE-EOM excitations.
+
+        Args:
+            excitations: EE-EOM excitations.
+
+        Returns:
+            EE-EOM excitations as a vector.
+        """
+        pass
+
+    @abstractmethod
+    def vector_to_excitations_ip(self, vector: NDArray[float]) -> tuple[Namespace[AmplitudeType], ...]:
+        """Construct a namespace of IP-EOM excitations from a vector.
+
+        Args:
+            vector: IP-EOM excitations as a vector.
+
+        Returns:
+            IP-EOM excitations.
+        """
+        pass
+
+    @abstractmethod
+    def vector_to_excitations_ea(self, vector: NDArray[float]) -> tuple[Namespace[AmplitudeType], ...]:
+        """Construct a namespace of EA-EOM excitations from a vector.
+
+        Args:
+            vector: EA-EOM excitations as a vector.
+
+        Returns:
+            EA-EOM excitations.
+        """
+        pass
+
+    @abstractmethod
+    def vector_to_excitations_ee(self, vector: NDArray[float]) -> tuple[Namespace[AmplitudeType], ...]:
+        """Construct a namespace of EE-EOM excitations from a vector.
+
+        Args:
+            vector: EE-EOM excitations as a vector.
+
+        Returns:
+            EE-EOM excitations.
+        """
         pass
 
     @property
@@ -230,12 +1054,12 @@ class BaseEBCC(ABC):
         """Initialise the fermionic space.
 
         Returns:
-            Fermionic space.
+            Fermionic space. All fermionic degrees of freedom are assumed to be correlated.
         """
         pass
 
     @abstractmethod
-    def get_fock(self) -> Any:
+    def get_fock(self) -> BaseFock:
         """Get the Fock matrix.
 
         Returns:
@@ -244,7 +1068,7 @@ class BaseEBCC(ABC):
         pass
 
     @abstractmethod
-    def get_eris(self, eris: Optional[Union[type[BaseERIs], NDArray[float]]]) -> Any:
+    def get_eris(self, eris: Optional[ERIsInputType] = None) -> BaseERIs:
         """Get the electron repulsion integrals.
 
         Args:
@@ -256,7 +1080,7 @@ class BaseEBCC(ABC):
         pass
 
     @abstractmethod
-    def get_g(self, g: NDArray[float]) -> Any:
+    def get_g(self, g: NDArray[float]) -> Namespace[Any]:
         """Get the blocks of the electron-boson coupling matrix.
 
         This matrix corresponds to the bosonic annihilation operator.
@@ -278,6 +1102,28 @@ class BaseEBCC(ABC):
         """
         pass
 
+    @property
+    def bare_fock(self) -> Any:
+        """Get the mean-field Fock matrix in the MO basis, including frozen parts.
+
+        Returns an array and not a `BaseFock` object.
+
+        Returns:
+            Mean-field Fock matrix.
+        """
+        pass
+
+    @property
+    @abstractmethod
+    def xi(self) -> NDArray[float]:
+        """Get the shift in the bosonic operators to diagonalise the photon Hamiltonian.
+
+        Returns:
+            Shift in the bosonic operators.
+        """
+        pass
+
+    @property
     def const(self) -> float:
         """Get the shift in energy from moving to the polaritonic basis.
 
@@ -287,16 +1133,6 @@ class BaseEBCC(ABC):
         if self.options.shift:
             return util.einsum("I,I->", self.omega, self.xi**2)
         return 0.0
-
-    @abstractmethod
-    @property
-    def xi(self) -> NDArray[float]:
-        """Get the shift in the bosonic operators to diagonalise the photon Hamiltonian.
-
-        Returns:
-            Shift in the bosonic operators.
-        """
-        pass
 
     @property
     def mo_coeff(self) -> NDArray[float]:
@@ -320,8 +1156,8 @@ class BaseEBCC(ABC):
             return np.asarray(self.mf.mo_occ).astype(types[float])
         return self._mo_occ
 
-    @abstractmethod
     @property
+    @abstractmethod
     def nmo(self) -> Any:
         """Get the number of molecular orbitals.
 
@@ -330,8 +1166,8 @@ class BaseEBCC(ABC):
         """
         pass
 
-    @abstractmethod
     @property
+    @abstractmethod
     def nocc(self) -> Any:
         """Get the number of occupied molecular orbitals.
 
@@ -340,8 +1176,8 @@ class BaseEBCC(ABC):
         """
         pass
 
-    @abstractmethod
     @property
+    @abstractmethod
     def nvir(self) -> Any:
         """Get the number of virtual molecular orbitals.
 
@@ -359,7 +1195,7 @@ class BaseEBCC(ABC):
         """
         if self.omega is None:
             return 0
-        return self.omega.shape[0]
+        return int(self.omega.shape[0])
 
     @property
     def e_tot(self) -> float:
@@ -368,7 +1204,7 @@ class BaseEBCC(ABC):
         Returns:
             Total energy.
         """
-        return types[float](self.mf.e_tot) + self.e_corr
+        return cast(self.mf.e_tot + self.e_corr, float)
 
     @property
     def t1(self) -> Any:
