+++ pyproject.toml
@@ -18,10 +18,10 @@ classifiers = [
         "Topic :: Scientific/Engineering",
         "Topic :: Software Development",
         "Programming Language :: Python",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
         "Operating System :: MacOS :: MacOS X",
         "Operating System :: POSIX :: Linux",
 ]
@@ -46,6 +46,7 @@ dev = [
     "flake8-pyproject>=1.2.3",
     "flake8-bugbear>=23.0.0",
     "flake8-docstrings>=1.6.0",
+    "mypy>=1.8.0",
     "coverage[toml]>=5.5.0",
     "pytest>=6.2.4",
     "pytest-cov>=4.0.0",
@@ -61,7 +62,7 @@ include = "ebcc"
 exclude = """
 /(
     | ebcc/codegen
-    | ebcc/__pycache__
+    | __pycache__
     | .git
 )/
 """
@@ -74,7 +75,7 @@ src_paths = [
     "ebcc",
 ]
 skip_glob = [
-    "ebcc/__pycache__/*",
+    "*/__pycache__/*",
     "ebcc/codegen/*",
     "ebcc/__init__.py",
 ]
@@ -101,10 +102,37 @@ include = "ebcc"
 exclude = """
 /(
     | ebcc/codegen
-    | ebcc/__pycache__
+    | __pycache__
     | .git
 )/
 """
+
+[tool.mypy]
+python_version = "3.10"
+exclude = """
+/(
+    | ebcc/codegen
+    | __pycache__
+    | .git
+)/
+"""
+
+[[tool.mypy.overrides]]
+module = "ebcc.*"
+strict = true
+ignore_missing_imports = true
+#implicit_optional = true
+disallow_subclassing_any = false
+implicit_reexport = true
+warn_unused_ignores = false
+
+[[tool.mypy.overrides]]
+module = "scipy.*"
+ignore_missing_imports = true
+
+[[tool.mypy.overrides]]
+module = "h5py.*"
+ignore_missing_imports = true
 
 [tool.coverage.run]
 branch = true
